Move auth out of dashbaord folder
