+++ src/App.js
@@ -9,7 +9,7 @@ import GetGroups from './components/redux/Groups/GetGroups';
 import Reservations from './components/Reservations';
 import NewReservation from './components/Reservations/NewReservation';
 import { sessionIsLoggedIn } from './redux/authentication/authentication';
-import Registration from './pages/dashboard/auth/Registartion';
+import Registration from './pages/auth/Registartion';
 
 function App() {
   const dispatch = useDispatch();
